+++ projects/mvc-with-jpa-1/src/main/java/com/example/mvc_with_jpa_1/repository/CommentEmRepository.java
@@ -37,7 +37,7 @@ public class CommentEmRepository {
 //                on p1_0.id=c1_0.post_id
 //        limit
 //            ?, ?
-    public List<Comment> findAllJoinFetchLimitByEm(int postId, Pageable pageable) {
+    public List<Comment> findAllJoinFetchLimitByEm(Long postId, Pageable pageable) {
 //        https://velog.io/@j3beom/JPA-JPQL-Fetch-Join
 //        https://www.inflearn.com/questions/15876/fetch-join-%EC%8B%9C-%EB%B3%84%EC%B9%AD%EA%B4%80%EB%A0%A8-%EC%A7%88%EB%AC%B8%EC%9E%85%EB%8B%88%EB%8B%A4
 //        fetch join 대상에는 별칭을 줄 수 없으나 hibernate 에서만 가능
